Add --log-handle flag
This allows winc to log to an arbitrary file handle. In order to use
this flag, the parent process must be sure that winc inherits the handle
as well.

The use case that this enables is for winc to log to a pipe, that the
parent can than read as data is available

[#154291838]

Signed-off-by: Sam Smith <sesmith177@gmail.com>