Increase ctrl+break test timeout.
- We actually need to change this test or change the implementation.
This test isn't really guaranteeing that the winc exec command is being
interrupted and exiting immediately. As we increased the graceful
shutdown time, it eventually became larger than the Eventually test
timeouts of 10 seconds. Q: why are stdout and stderr still blocking in
this test? Q: did this test ever really work or was it a false positive?

[#166002871]

Signed-off-by: Genevieve LEsperance <glesperance@pivotal.io>