Remove the CTRL+C tests
After making the graceful shutdown changes, the process stops
after ~11 seconds.
The tests that were deleted expected the process to be killed after 12 seconds but it was
flaky due to the 1 second difference.
Before the graceful shutdown implementation, the process was killed immediately and the test expected
it to be killed after 10 seconds.

Will return this test if we'll decide to handle CTRL+C.

[#166340703]

Signed-off-by: Arjun Sreedharan <asreedharan@pivotal.io>